requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
SQLAlchemy==2.0.28
Werkzeug==3.0.1
email-validator==2.1.1
python-dotenv==1.0.1 